+++ src/Core/Spec/Spec13.php
@@ -111,4 +111,9 @@ final class Spec13 implements SpecInterface
     {
         return true;
     }
+
+    public function requiresComponentVersion(): bool
+    {
+        return true;
+    }
 }